feat: support JS-only transforms
For TSX files, compile first to JSX if any JSX-only transforms are present. Then compile to JS if any JS-only transforms are present.

For TS and JSX files, compile to JS if any JS-only transforms are present.

Since `esbuild.transform` doesn't access the filesystem, we have to load the tsconfig manually (using the "tsconfck" package).